tests for the patterns implemented so far

+++ src/main/java/com/selzerj/geopattern/composers/PatternPreset.java
@@ -13,8 +13,8 @@ public class PatternPreset {
 	Color fillColorLight;
 
 	Color strokeColor;
-	float strokeOpacity;
+	double strokeOpacity;
 
-	float opacityMin;
-	float opacityMax;
+	double opacityMin;
+	double opacityMax;
 }
