+++ src/main/java/com/selzerj/geopattern/svg/SvgImage.java
@@ -31,7 +31,7 @@ public class SvgImage {
 		return getSvgHeader() + body + getSvgCloser();
 	}
 
-	public SvgImage addRect(float x, float y, String width, String height, Map<String, String> args) {
+	public SvgImage addRect(double x, double y, String width, String height, Map<String, String> args) {
 		body += String.format("<rect x=\"%f\" y=\"%f\" width=\"%s\" height=\"%s\" %s />",
 				x, y, width, height, writeArgs(args));
 		return this;
